requirements.txt
flask == 3.0.3
python-dotenv == 1.0.1
openai == 1.45.0
aiohttp == 3.10.5
requests == 2.32.3
pymysql == 1.1.1
sqlalchemy == 2.0.35
flask_sqlalchemy == 3.1.1